Don't create tarball attribute with confusing name
The cbspkgs attribute contains a derivation that returns this
repository as a tarball. This is not useful, because the repository is
public and can be fetched just as well via git directly or via the
"Download as ZIP" support of Github.

The attribute was also shadowed by the similar attribute in our
internal overlay.